+++ src/stories/changelog/changelog-data.ts
@@ -11,6 +11,24 @@ export type ChangeItem = {
 
 export const changelogData: ChangeItem[] = [
   {
+    version: "1.8.2",
+    author: "Gabriel Glatz",
+    date: "2026-01-16",
+    shortTitle: "Chart",
+    added: [
+      "Chart: adiconado tela de Loading e Ajustado label em ocassioes de barras muito pequenas",
+      "Tooltip: adiconado novo comportamento",
+    ],  
+    changed: [
+      "Selects: refatorado sistema de estilos e ícones para manter consistência visual entre Select, Combobox e MultiSelect.",
+      "Date Time Picker: refatorado sistema de estilos e ícones para manter consistência visual entre Select, Combobox e MultiSelect.",
+    ],
+    fixed: [
+      "Selects: corrigido bug crítico do formatador de legenda que não aplicava corretamente a função customizada de formatação, resultando em exibição incorreta de labels no eixo e tooltips interativos, afetando principalmente gráficos com valores monetários e percentuais.",
+      "Selects: corrigido bug crítico do formatador de legenda que não aplicava corretamente a função customizada de formatação, resultando em exibição incorreta de labels no eixo e tooltips interativos, afetando principalmente gráficos com valores monetários e percentuais.",
+    ],
+  },
+  {
     version: "1.8.1",
     author: "Gabriel Glatz",
     date: "2026-01-14",